+++ src/lib/components/PureDecrement.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 /**
- * @typedef {Object} PureIncrement
+ * @typedef {Object} PureDecrementProps
  * @property {number} start - The starting count.
  * @property {number} end - The ending count.
  * @property {number} duration - The duration of the decrement animation.
@@ -10,22 +10,27 @@ import React, { useState, useEffect } from 'react';
 /**
  * PureDecrement component.
  *
- * @param {PureIncrement} props - The component props.
+ * @param {PureDecrementProps} props - The component props.
  * @returns {JSX.Element} - The rendered JSX element.
  */
-function PureIncrement({ start, end, duration, className }) {
+
+
+interface PureDecrementProps {
+  start: number;
+  end: number;
+  duration: number;
+  className?: string;
+}
+const PureDecrement: React.FC<PureDecrementProps> = ({ start, end, duration, className }) => {
   const [count, setCount] = useState(start);
 
   useEffect(() => {
-    const increment = Math.ceil((end - start) / (duration * 60));
-    const timeout = Math.ceil((duration * 1000) / (end - start));
+    const decrement = Math.ceil((start - end) / (duration * 60));
+    const timeout = Math.ceil((duration * 1000) / (start - end));
 
-    if (start < end) {
+    if (start > end) {
       const timer = setInterval(() => {
-        setCount((prevCount) => {
-          const newCount = prevCount + increment;
-          return newCount >= end ? end : newCount;
-        });
+        setCount((prevCount) => Math.max(prevCount - decrement, end));
       }, timeout);
 
       return () => clearInterval(timer);
@@ -35,6 +40,6 @@ function PureIncrement({ start, end, duration, className }) {
   }, [start, end, duration]);
 
   return <span className={className}>{count}</span>;
-}
+};
 
-export default PureIncrement;
+export default PureDecrement;
