--- src/lib/index.js
@@ -1,3 +1,0 @@
-import PureIncrement from './components/PureIncrement'; // Import PureIncrement
-import PureDecrement from './components/PureDecrement'; // Import PureDecrement
-export { PureIncrement, PureDecrement };
