+++ src/lib/index.tsx
@@ -1,0 +1,2 @@
+export { default as PureIncrement } from './components/PureIncrement';
+export { default as PureDecrement } from './components/PureDecrement';
